+++ prime-router/src/main/kotlin/history/azure/DatabaseDeliveryAccess.kt
@@ -3,6 +3,7 @@ package gov.cdc.prime.router.history.azure
 import gov.cdc.prime.router.ReportId
 import gov.cdc.prime.router.Topic
 import gov.cdc.prime.router.azure.DatabaseAccess
+import gov.cdc.prime.router.azure.db.Tables
 import gov.cdc.prime.router.azure.db.Tables.ACTION
 import gov.cdc.prime.router.azure.db.Tables.REPORT_FACILITIES
 import gov.cdc.prime.router.azure.db.Tables.REPORT_FILE
@@ -51,8 +52,8 @@ class DatabaseDeliveryAccess(db: DatabaseAccess = BaseEngine.databaseAccessSingl
                         .and(REPORT_FILE.TRANSPORT_PARAMS.isNotNull)
                         .and(REPORT_FILE.TRANSPORT_RESULT.notLike("%downloadedBy%"))
                         .and(REPORT_FILE.SCHEMA_TOPIC.equalIgnoreCase(Topic.ELR_ELIMS.jsonVal))
-            )
-            .and(REPORT_FILE.RECEIVING_ORG.eq(organization))
+                )
+                .and(REPORT_FILE.RECEIVING_ORG.eq(organization))
         } else {
             REPORT_FILE.NEXT_ACTION.eq(TaskAction.send)
                 .and(REPORT_FILE.RECEIVING_ORG.eq(organization))
@@ -62,6 +63,25 @@ class DatabaseDeliveryAccess(db: DatabaseAccess = BaseEngine.databaseAccessSingl
             filter = filter.and(REPORT_FILE.RECEIVING_ORG_SVC.eq(orgService))
         }
 
+        return filter
+    }
+
+    /**
+     * Filter for sent reports for a particular receiver organization.
+     * @param organization is the Organization Name returned from the Okta JWT Claim.
+     * @param orgService is a specifier for an organization, such as the client or service used to send/receive
+     */
+    fun deliveredReportsByOrgFilter(
+        organization: String,
+        orgService: String?,
+    ): Condition {
+        var filter = Tables.REPORT_FILE.NEXT_ACTION.isNull.and(Tables.REPORT_FILE.TRANSPORT_PARAMS.isNotNull)
+            .and(Tables.REPORT_FILE.DOWNLOADED_BY.isNull)
+            .and(Tables.REPORT_FILE.RECEIVING_ORG.eq(organization))
+
+        if (orgService != null) {
+            filter = filter.and(Tables.REPORT_FILE.RECEIVING_ORG_SVC.eq(orgService))
+        }
         return filter
     }
 
@@ -78,31 +98,31 @@ class DatabaseDeliveryAccess(db: DatabaseAccess = BaseEngine.databaseAccessSingl
         orgName: String?,
         klass: Class<T>,
     ): T? = db.transactReturning { txn ->
-            DSL.using(txn)
-                .select(
-                    ACTION.ACTION_ID,
-                    ACTION.CREATED_AT,
-                    ACTION.SENDING_ORG,
-                    REPORT_FILE.RECEIVING_ORG,
-                    REPORT_FILE.RECEIVING_ORG_SVC,
-                    ACTION.HTTP_STATUS,
-                    ACTION.EXTERNAL_NAME,
-                    REPORT_FILE.REPORT_ID,
-                    REPORT_FILE.SCHEMA_TOPIC,
-                    REPORT_FILE.ITEM_COUNT,
-                    REPORT_FILE.BODY_URL,
-                    REPORT_FILE.SCHEMA_NAME,
-                    REPORT_FILE.BODY_FORMAT
+        DSL.using(txn)
+            .select(
+                ACTION.ACTION_ID,
+                ACTION.CREATED_AT,
+                ACTION.SENDING_ORG,
+                REPORT_FILE.RECEIVING_ORG,
+                REPORT_FILE.RECEIVING_ORG_SVC,
+                ACTION.HTTP_STATUS,
+                ACTION.EXTERNAL_NAME,
+                REPORT_FILE.REPORT_ID,
+                REPORT_FILE.SCHEMA_TOPIC,
+                REPORT_FILE.ITEM_COUNT,
+                REPORT_FILE.BODY_URL,
+                REPORT_FILE.SCHEMA_NAME,
+                REPORT_FILE.BODY_FORMAT
+            )
+            .from(
+                ACTION.join(REPORT_FILE).on(
+                    REPORT_FILE.ACTION_ID.eq(ACTION.ACTION_ID)
                 )
-                .from(
-                    ACTION.join(REPORT_FILE).on(
-                        REPORT_FILE.ACTION_ID.eq(ACTION.ACTION_ID)
-                    )
-                )
-                .where(
-                    ACTION.ACTION_ID.eq(actionId)
-                ).fetchOne()?.into(klass)
-        }
+            )
+            .where(
+                ACTION.ACTION_ID.eq(actionId)
+            ).fetchOne()?.into(klass)
+    }
 
     override fun <T> fetchRelatedActions(reportId: UUID, klass: Class<T>): List<T> {
         TODO("Not yet implemented")
